refactor: move to v1 namespace
