+++ app/Http/Controllers/Api/V1/AuditLogController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\V1;
 
 use App\Enums\AuditLogAction;
 use App\Http\Controllers\Controller;
